# Discussion: switch minio/blake2b-simd to golang.org/x/crypto